Creation of transaction with security chaecks

+++ project/paybar-clearingHouse/clearinghouse-ejb/src/main/java/paybar/model/DetailAccount.java
@@ -7,7 +7,6 @@ import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
@@ -184,7 +183,7 @@ public class DetailAccount implements Serializable {
 		this.coupons = coupons;
 	}
 
-	public float getCredit() {
+	public long getCredit() {
 		return credit;
 	}
